motor noise testing added with additions to common for testing cli's and changes to the calibration cli, began adding filters to accelerometer data for testing also

+++ common/lib/MotorControlManager/MotorControlManager.h
@@ -18,7 +18,7 @@ public:
     digitalWriteFast(pin, LOW);
   }
 
-  void trigger(int pulseTime) {
+  void trigger(float32_t pulseTime) {
     on();
     TeensyDelay::trigger(pulseTime, channelId);
   }
